requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
icrawler>=0.6.6
wandb>=0.15.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.65.0
pyyaml>=6.0
kaggle>=1.5.0
